Add concurrently and babel packages to backend
The concurrently package was added as a dev dependency to aid development.
Instead of starting the back and front ends on separate terminals,
a node script could be executed to run both within one terminal session.

The babel package was added due to a warning message when running the
frontend. At this time, create-react-app is not maintained anymore so
there are many warnings about deprecated dependencies.