Add GET route to provide requested page info
Closes #3

The GET route is housed in a controller file and utilizes helpers
to flatten and unflatten a given JSON object. This is to make the process
of replacing image paths (public IDs in Cloudinary) with Cloudinary
URLs. The URLs are generated using a helper as well.

The homepage data was modified to rename a key for an svg file to have
it align with the other image key names and make the process of converting
the path to a URL easier.