Refactor GET route to serve header and footer data
Closes #7

The GET route in pageController.js was simplified by moving the
readFile code out into a helper function and having the readFile
helper function return a promise of the file data.

The page data is served as an array in the following order: header,
page content, and footer to match the flow/layout of the website.