Update homepage, header, footer JSON files
Closes #8

The files were updated to add alt and aria children properties to each
image and icon property. These properties will contain descriptions of the
images and icons for accessability reasons. Images and icons that are
reused at different breakpoints all share the same alt or aria
descriptions to prevent data duplication.

In addition, "path" properties were added for images and icons to simplify the
sort process when replacing the image and icon paths with Cloudinary
URLs.