Refactor to request header and footer data once
Closes #11

This refactor moved router configuration from App.js to index.js and
set App.js as the root route. Doing so allows the header and footer
data to be only requested once, stored in a higher level component,
and passed down/shared using context.

The pageRouter was refactored so that only one file can be requested at a
time and the header and footer data isn't always provided with a request
for data.

Utils and helper functions were updated/created to function with the
changes above.