Create about.json and add to loader callback
Closes #14

about.json was created to house all the data from the previous
property "closing" since the data is reused across the whole website.
The data in the file will be fetched only once like header.json and
footer.json to reduce requests to the backend. The data is available
through context as well.