Style Banner Section and fix Header Section
Closes #21

The custom theme file was updated to fix the missing/ mistyped style
attributes, rename custom style names to better indicate where the style
comes from, and extend the Text component for the New Product component.

As part of this commit, elements of the Banner component were turned into
React components to follow React convention.

The Header section was updated as part of this commit as well to fix styling
issues that were not seen due to inaccurate color display of my laptop
monitor.