Style Product Category Section of Home pages
Closes #22

As part of this commit, new reuseable, components were created for the
thumbnails to simplify the styling process. The Earphone thumbnail was the
only one that needed a slightly different styling adjustment compared
to the other thumnbails due to the positioning of the earphone in the
provided image.

Index.js was updated to remove an unused import.