Style product cards section of homepage
Closes #23

Part of this commit also includes updates to the Chakra UI
custom Button and colors files. A new button variant was created for the
primary product card and a new color was added for the new button
variant. The new button variant was not captured in the provided
design system.