Styled footer section of home page for mobile view
Closes #25

ReactSVG package was added to inject the social media svg icons into the
page and simplify styling it when the icon is hovered over. Originally,
the svg icons were displayed using an img tag. Changing the color of the
icons with the img tag proved difficult since the img tag does not allow
you to directly change the color of the path tag using css.

For the time being, an index.css file was created to house a class that allows
directly changing the icon color when hovered over. Attempting to change it
through Chakra UI did not work.