Organize files; create new and restyle components
The files were organized into folders based on the section of the
page they were used on. Components that are used in multiple sections
of the page do not belong to a specific page section folder.

index.css was updated to generalize the name so that it could be used
on any icon.