Refactor styles and create new components
Closes #26

Some inline text styles were moved into a new theme file for
Chakra UI to minimize code duplication for text components. Future
text styles will be added to the new theme file.

Due to some bugs with Chakra UI (see textstyles.js or headingStyles.js)
heading components will be housed in a object not based on Chakra UI's
theme structure. They will be applied inline as needed for heading components.

New components for the headline in the closing section of the homepage
were created since it's cleaner to have the code within the component instead
of storing it in a helper function.