Create tests for homepage and refactor for tests
Closes #28

New Cypress tests were created to verify that the content from the backend
was being displayed/ served correctly. Many components were updated to
add an id attribute to simplify the testing process.

The header and footer sections are housed in separate files to avoid duplicating
the tests across the future pages that will be created for the website.

In addition, eslint was added to remove the errors for cypress syntax and
and lint the rest of the code for style consistency.