Add hamburger menu button functionality for header
Closes #30

A new overlay component was created to display the navigation links for the
mobile and tablet views when the hamburger menu is clicked on.

New react contexts were added to handle and house the data for the
menu overlay styles since it'll be reused on all web pages. The context
for menu will be shared to <Outlet /> whenever a new page is rendered.

Fixed a bug for the background for the thumbnails not appearing appearing.