Create and style new headphone category page
Closes #33

A new headphone category page was created and linked to applicable
url redirect associated components on the homepage.

Many shared components were restyled to make them work on
both the home and headphone category pages. The styling mainly
consisted of spacing spacing changes with some fixes to background
color or opacity.

A new textStyle property was created primarily for the
product preview description.

A new css class was created to change the shop link text color due
to using a third party component and needing to target child components
not available to me.