Create Speakers page and refactor to make template
Closes #35

The Speakers category page was created and relevant components were updated
to redirect to the new page. A new JSON file was created for the Speakers page
to store the data that the backend will serve too.

A new base template was created for the product category pages since they
all share a similar page structure.

A few unused files were removed.

Thumbnails that were missing or had the wrong redirect URL prop were fixed.