Created Earphone page and refactor base template
Closes #36

A new Earphone category page was created and styled per the
Figma design. A new JSON file to house and serve all relevant data for
the page was created too.

The base product category page was updated to move the ProductCategoryHeading
component from the individual category pages to simplify the code and
keep shared components on the base template.