Create product detail pages and update template
Closes #42

The following product detail pages were created:
- XX99 Mark I
- XX59
- ZX9
- ZX7
- YX1 Wireless

For each page, a new JSON file was created with the relevant data.

The base template was also slightly updated to fix an issue where the
product names were not breaking at the right spot.

The XX99MarkII JSON file was updated to clarify the name of the route properties
and match the other JSON files.