Style Home page for tablet and desktop layouts
Closes #46

Most components used on the Home page were styled for the tablet and
desktop views.

In addition, the closing, footer, and heading JSON files were updated to
fix some typos and/or simplify the key-value structure.

The LineBreak component was deleted. The Header container was restyled
so that the line break is displayed through the bottom border of the container.