Add tablet & desktop layout to product cat. pages
Closes #48

The base template product category page was updated to add tablet and desktop
layouts per the Figma design.

The template was also updated to procedurally render the page since all the components
on the page can be reused with different product content.

The Closing and Header sections of website were also updated to fix spacing
issues and match the Figma wire frame designs.

index.js was updated to render the base template instead of the page-specific
components. As a result of using the base template, the page-specific
components were deleted since they will no longer be used.