Add Cart Modal to website
Closes #50

A cart modal was added to the website to display all products
a user plans on purchasing. The cart show the product image, product
price, quantity, and total price. The cart currently has dummy variables
and content for the time being until a feature is implemented to track and
allow a user to add products to the cart.

QuantityCounter was also moved from the productDetail folder to the shared
folder since the component is now used on multiple areas of the website.

header.json was updated to add images and prices related to the cart.