Created Checkout Page with Form Validation
Close #52

The checkout page was styled for the mobile, tablet, and desktop
views. Functionality of the go back button was not added.

Validation for the checkout form primarily involves preventing typos and
minor user input mistakes. Validation to determine if the information
given is real was not included since it is outside the scope of this project
and will be difficult to determine. 3rd party API exists to help determine that
but will not completely solve the issue.

The React Hook Form package was installed on the front end to aid structuring the
form and for basic validation. It will also help with submitting the data
to the database in a future feature update.

A new checkout JSON file was also added to house data specifically for the
checkout page.

New Chakra UI themes and components were created for the checkout page. The
input component style file was deleted and moved in a theme folder due to
it being a multipart component.