Create checkout order confirmation modal
Closes #53

The modal displays one item that the user purchased. If more than one
item was purchased, an option is provided for the user to display all
the purchased items. It also displays the grand total of the purchase.

Users are not allowed to close the modal unless the BackToHomeButton is
clicked on. If the user decides to go back to the previous page nothing is
performed to handle that. That situation will be handled in the future.

The modal was styled for the mobile, tablet, and desktop views.

The ContinueAndPayButton.js component was renamed to PayButton.js to
shorten the name.

A "size" prop was added to the CartItemImage component to change the size
of the product image and allow it to be reusable.