Refactor Tablet and Mobile Menu into Modal
Closes #54

The custom menu modal I created was refactored using
Chakra UI's Modal components. This was done to standardize
the code and simplify development.

The refactor allowed me to remove unneeded files and
event handlers; letting Chakra UI handle it behind the
scene.

Functionality was also added to handle how the modal was
closed when the logo or cart is clicked on.