Fix checkout and cart functionality bugs
Fixes #55

Users are no longer able to go to the checkout page if the cart is empty.
The checkout button is disabled if the cart is empty.

On the checkout page, clicking the pay button does not complete the purchase
if the cart is empty or if the form has errors. This required lifting the
react hook form API components up to CheckoutPage.js so that relevant data
and functions are shared between react child components.

PayButton was also lifted up from CheckoutPriceSummary.js to
CheckoutSummary.js since it has nothing to do with displaying prices.

CheckoutItems() in CheckoutSummary.js was updated to fix a bug where the
length of the returned array value did not reflect how many items were
actually in the cart.

The radio inputs were updated to register them to react hook form. The e-money
input fields were updated so that the values from the fields are conditionally
submitted with the form data depending on the radio option clicked. A bug
was also fixed where the e-money input fields were preventing the form from
being submitted if they had any errors. This required resetting the
error state of the fields.