Fix Cart modal moving past marked point in Figma
Closes #58

Per the Figma wire frame, the cart modal in the tablet and desktop views
should not move part the cart icon. This was fixed by adding an additional
container to limit the movement space. Adding the extra container
required setting an event listener for <ModalContent> to make the modal
close on click outside the visible portions of the cart. onClick functions
were also added to the <ModalContent> children to prevent event
propagation from triggering the event listener in <ModalContent>
to close the modal.