Fix Cart Modal overflow and scrollbar UI issues
Fixes #59

The scrollbar spacing issue required the use of the MutationObserver API
to detect if the <ModalBody> component exists in the DOM. If it exists
(when the modal is loaded), then right padding is added to <ModalBody>
to space the cart items and scrollbar.

The observer is only active if the cart icon is clicked on.