Connect backend to MongoDB to store purchase order
Closes #61

The express-async-errors package was installed to simplify error handling.
Errors are automatically passed onto error handling middleware and does not
require the use of the next() method.

Backend form validation was updated to add a missing field (radio field)
and fix a logical issue where the fields eMoneyNumber and eMoneyPin are
validated based on the value of the radio field.

A new document schema purchaseOrder was created and MongoDB was updated
accordingly to store the purchaseOrder documents in the audiophile database.