Added products and quantity to purchase order data
Closes #62

The data sent to the backend was updated to add data involving the
products purchased and the quantity of each purchased product.

Doing so required changing the content-type of the requests from
x-www-form-urlencoded to application/json. This was performed to simplify
the process of handling the data since the new data being added required
converting it from an object into x-www-form-urlencoded compliant formatting.