Add feature to calc and store purchase order costs
Closes #63

The following costs are broken down and included to the document schema:
total (before vat and shipping), shipping, vat (not including shipping),
and grand total (total + vat + shipping).

The validation array was moved into a separate, new file in the utilities
folder to help with code organization and maintainability.

A new json file was created specifically for the product prices to aid with
backend cost calculations.