Fix cannot update component while rendering error
Fixes #66

The cause of the error was the use of the setGrandTotal state function
in App.js. THe calcGrandTotal function in the CheckoutSummary component
was changing the state of the grandTotal state variable while rendering
the CheckoutSummary component. As a result, the variable was removed and
a different approach was taken to calculate the costs and pass the values
to different components. Costs are now calculated at the App.js level
and passed down to child components.

All cost calculating functions were moved into the helper utility file
to reduce code duplication and remove the need for the grandTotal state
variable. The checkoutItems function was refactored and moved into the
helper utility file to reduce code duplication as well. numToLocaleStr was
created to create boilerplate code.