Create 404 Page when user on non-existent page
Closes #68

A new error page was created and styled per the figma wire frame
designs.

The router in index.js was moved into a new file to clean
up the code.

Eslint configuration file was updated to change the ecmaVersion from 2018
to 'latest' due to it not recognizing the nullish coalescing operator.

Error elements within router were added at the root and child levels in
order to get <ErrorPage/> to render in place of <Outlet/> and maintain
the website page layout. Following the method in react router's tutorial
did not work. See https://stackoverflow.com/questions/75179651/how-can-i-make-a-pathless-route-render-correctly-inside-outlet-in-react-router for the
solution.