Add backend feature to send order confirm emails
Closes #70

The following dependencies were added to have the backend send emails:
- Resend: Email API and vendor for sending emails
- React Email: Used to create the email template
- Babel: Used to transpile .jsx files and ES6 code into javascript that
that can be interpreted by Nodejs.
- React: Used to create the email template

The React Email and Resend packages were used to help simplify the creation
and sending of emails. I'm familiar with the React framework library and
Resend handles the email configurations needed to email users of their
purchase order.

In order to use React Email, Babel was used to transpile the .jsx files
used to build the email template into a readable format for NodeJs.
The email sending utility needed to be converted as well. They must be
compiled everything a change is made before running the application.

Minor bug(s) fixed on the frontend include the following:
- Fixed checkout summary costs not displaying 2 decimal places.
- Fixed product counter on product detail pages carrying over previous
number when moving to another product detail page by using the "see product"
button. setState for counter was lifted up and setState function passed
to "see product" button to have it reset counter to 1 on click.

Minor bug(s) fixed on backend include the following:
- fixing display name of XX99 MARK II and MARK I products from "MARK"
to "MK" to match figma design.