Add toast and update cart to show items in cart
Closes #72

Added a toast notification with a countdown time of 5 seconds to indicate
to the user an item was added to their cart. The user has the ability to
manually close the toast if needed.

An exclamation point button was added next to the cart icon to help
indicate to the user there are items in their cart. It is hidden when
the cart is empty. It was made into a button to give the user the option of
clicking on it to view their cart.