Create cart and checkout demo buttons
Closes #73

Two demo buttons were created: a cart and checkout demo button. The
cart demo button only adds items to the cart whereas the checkout demon
button only fills out the checkout form. These buttons are only used to
demonstrate the functionality of the website.

The demo buttons were styled differently from the website to help indicate
that they are not part of the original design provided by frontend mentor.

The cart modal was restyled to create space for the demo button and fix
an issue where the "XX99 MK II" product name was wrapping due to a lack
of space.

The cart demo and exclamation buttons were updated to add a listener to
pad the cart line items when a scrollbar appeared. The functions responsible
for spacing the line items were moved into the helper.js utility file to
reduce code duplication.