Deploy website to internet
Closes #75

Fly.io was used as the hosting vendor primarily because they offer a
generous "hobby" plan that will allow me to host the website for free.

Fly.io dependency packages were installed to deploy the website. NodeJS and
npm were also updated to the latest LTS versions.

Dev dependencies and environmental variables were commented out and added,
respectively prior to deploying the website.

A UI bug for the mobile menu on the frontend was fixed.