Fix error on non-home page refresh or url location
Fixes #78

The cause of the bug is due to not having server-side routing.

When a page is refreshed, manually typed in,
or shared with other people, the browser makes a GET request to the
server at the intended URL route. When this happens, the server has no way
of handling the route and throws an error.

The fix involves adding a middleware to handle server-side routing. The
middleware (routeController) uses the catch-all method to direct all routes
to index.html. From there, React Router will handle the rest on the client side
and direct the user to the page they wanted.

On the frontend, the home page route was updated to replace the path for the
home page with "index:true" to set it as the default route.

A minor bug on the frontend where the wrong key name was used for the logo
home page route was fixed.