Add feature to track user's cart
Closes #79

The application tracks the user's cart by assigning a unique cart ID to
a user's browser and updating a cart document in the database every time
a change is made to the cart.

The cart ID is stored using the localStorage in a user's browser. LocalStorage
was chosen over cookies because:
- not every request requires a cookie for my application
- localStorage is the preferred/ modern way of storing data on a browser
client

A couple of page files were also renamed to clarify what they were and
comments were added/ improved to help document/ explain certain sections of
code.

A new helper function (addItemsToCart) was added to reduce code duplication.

The purchaseComplete field of the cart schema was added to avoid adding
purchased items back in the cart. The feature will be implemented in a
future issue.