Create new cart for user after completing purchase
Closes #80

After completing a purchase, the user's cart data is marked as completed
and saved for analysis (if this were a real company). A new cart is created for
the user and a new cart ID is assigned to the user's browser client to track
any future purchases.

As part of this update, code that creates the new cart was lifted into the
routerHelper file so it can be reused.

The static_build folder was also renamed to build to simplify the build
process.