Create domain and configure sending email
Closes #81

A new domain was purchased from Namecheap and configured for the
website: https://audiophile.fan.

The following sub-domains were created for the website so far: www and
purchase. The "purchase" sub-domain is only used for sending
confirmation emails to users to complete purchases at this time. It was
created to:
- separate the action of sending the email from the root domain
- security reasons (prevents DDOS attacks from taking the root domain down)
- clarify the sending purpose

Bugs with the email template were also fixed. The logo and website link
were updated to point to the new domain. Previously, it did not point to
any domains at all.